added .env/ adapted other files for .env
