+++ .gitignore
@@ -1,3 +1,6 @@
 # ignore pycache
 __pycache__/
 *.pyc
+
+# environments
+.env
